Change request / response to carry boolean arrays / short arrays Add DataType to help user to convert data from registers Add a simple sample to show how to use DataType

+++ src/test/java/de/gandev/modjn/test/ModbusWriteSingleRegisterTest.java
@@ -47,7 +47,7 @@ public class ModbusWriteSingleRegisterTest {
         boolean state = true;
         int value = 0x0000;
         for (int i = 0; i < 20; i++) {
-            WriteSingleRegister writeRegister = modbusClient.writeSingleRegister(12321, value);
+            WriteSingleRegister writeRegister = modbusClient.writeSingleRegister(12321, (short)value);
 
             assertNotNull(writeRegister);
